system/core: no longer poweroff when charger requests reboot
The patch was originally added under the assumption that the container
management will reboot the machine when the container powers off [1]
(and simply restart the container when it requests reboot).

However, on Ubuntu Touch Focal we don't do this anymore, as we have a
proper way to convey reboot and poweroff from container to the host [2].
So stop changing the charger binary's intention, which allow charger to
reboot to normal system when holding power button and to power off when
the cable is unplugged.

If we were still producing Ubuntu Touch Xenial image, we would have to
add a post-stop script similar to [2]. However since all Xenial image
pipelines are stopped, I think there's nothing else that we have to do.

[1] https://gitlab.com/ubports/development/core/rootfs-builder-debos/-/blob/master/android9/overlay/etc/init/charger-reboot.conf
[2] https://gitlab.com/ubports/development/core/hybris-support/lxc-android-config/-/blob/ubports/focal/var/lib/lxc/android/post-stop.sh

Change-Id: I908cd5e9a56c4c792559b38eb78e1f56823fcbca